Migrated to Truffle, added unit tests
Because Remix couldn't cope with the size anymore.

Unit tests involve tests on the ledger and ledger lib. Unfortunately,
TestRPC 4.0.1 has a bug that makes the MarketStake function 'deposit',
which calls a library function, take twice the message value. The ledger
seems unaffected, it only appears to affect the balance of the
MarketStake contract.